requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
onnxruntime==1.22.0
numpy==2.2.4
opencv-python-headless==4.11.0.86
supabase==2.15.1
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.11.1
